requirements.txt
fastapi
pydantic
pydantic-settings
aio-pika
uvicorn
python-multipart
pillow
